separate handlers.py and create logger
